unposted-audio-journal/server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai-whisper==20231117
torch==2.1.0
transformers==4.35.0
librosa==0.10.1
numpy==1.24.3
scipy==1.11.3
soundfile==0.12.1
